requirements.txt
aniso8601==9.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
demoji==1.1.0
Flask==2.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
nltk==3.7
numpy==1.22.3
pandas==1.4.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
regex==2022.4.24
requests==2.27.1
schedule==1.1.0
six==1.16.0
torch==1.11.0
tqdm==4.64.0
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
